feat: cascade changes in clarinet test
