+++ src/runnner/mod.rs
@@ -6,12 +6,100 @@
 #![allow(non_upper_case_globals)]
 #![allow(unused_must_use)]
 
-use clarity_repl::repl::Session;
+use crate::deployment::{
+    apply_on_chain_deployment, check_deployments, display_deployment, generate_default_deployment,
+    get_absolute_deployment_path, get_default_deployment_path, initiate_session_from_deployment,
+    load_deployment, read_or_default_to_generated_deployment, setup_session_with_deployment,
+    update_session_with_contracts_executions, update_session_with_genesis_accounts,
+    write_deployment,
+};
+use clarity_repl::clarity::analysis::contract_interface_builder::{
+    build_contract_interface, ContractInterface,
+};
+use clarity_repl::clarity::types::QualifiedContractIdentifier;
+use clarity_repl::repl::ast::ContractAST;
+use clarity_repl::repl::{ExecutionResult, Session};
 
-use crate::utils;
+use std::collections::{BTreeMap, HashMap};
 use std::path::PathBuf;
 
+use crate::deployment::types::DeploymentSpecification;
+
+pub mod api_v1;
+mod costs;
 pub mod deno;
+mod utils;
+
+#[derive(Clone)]
+pub struct DeploymentCache {
+    session: Session,
+    session_accounts_only: Session,
+    deployment_path: Option<String>,
+    deployment: DeploymentSpecification,
+    contracts_artifacts: HashMap<QualifiedContractIdentifier, AnalysisArtifacts>,
+}
+
+impl DeploymentCache {
+    pub fn new(
+        manifest_path: &PathBuf,
+        deployment: DeploymentSpecification,
+        deployment_path: &Option<String>,
+        mut contracts_asts: Option<HashMap<QualifiedContractIdentifier, ContractAST>>,
+    ) -> DeploymentCache {
+        let mut session_accounts_only = initiate_session_from_deployment(&manifest_path);
+        update_session_with_genesis_accounts(&mut session_accounts_only, &deployment);
+        let mut session = session_accounts_only.clone();
+
+        let contracts_asts = match contracts_asts.take() {
+            Some(asts) => asts,
+            None => HashMap::new(),
+        };
+
+        let execution_results = update_session_with_contracts_executions(
+            &mut session,
+            &deployment,
+            Some(contracts_asts),
+        );
+
+        let mut contracts_artifacts = HashMap::new();
+        for (contract_id, execution_result) in execution_results.into_iter() {
+            let mut execution_result = match execution_result {
+                Ok(execution_result) => execution_result,
+                Err(_) => {
+                    println!("Error found in contract {}", contract_id);
+                    std::process::exit(1);
+                }
+            };
+            if let Some((_, source, functions, ast, analysis)) = execution_result.contract.take() {
+                contracts_artifacts.insert(
+                    contract_id.clone(),
+                    AnalysisArtifacts {
+                        ast,
+                        interface: build_contract_interface(&analysis),
+                        source,
+                        dependencies: vec![],
+                    },
+                );
+            }
+        }
+
+        DeploymentCache {
+            session,
+            session_accounts_only,
+            deployment_path: deployment_path.clone(),
+            contracts_artifacts,
+            deployment,
+        }
+    }
+}
+
+#[derive(Clone)]
+pub struct AnalysisArtifacts {
+    pub ast: ContractAST,
+    pub interface: ContractInterface,
+    pub dependencies: Vec<String>,
+    pub source: String,
+}
 
 pub fn run_scripts(
     files: Vec<String>,
@@ -21,7 +109,7 @@ pub fn run_scripts(
     allow_wallets: bool,
     allow_disk_write: bool,
     manifest_path: PathBuf,
-    session: Option<Session>,
+    cache: DeploymentCache,
 ) -> Result<u32, (String, u32)> {
     match block_on(deno::do_run_scripts(
         files,
@@ -31,7 +119,7 @@ pub fn run_scripts(
         allow_wallets,
         allow_disk_write,
         manifest_path,
-        session,
+        cache,
     )) {
         Err(e) => Err((format!("{:?}", e), 0)),
         Ok(res) => Ok(res),
@@ -42,42 +130,6 @@ pub fn block_on<F, R>(future: F) -> R
 where
     F: std::future::Future<Output = R>,
 {
-    let rt = utils::create_basic_runtime();
+    let rt = crate::utils::create_basic_runtime();
     rt.block_on(future)
 }
-
-// struct ClaritestTransaction {
-// }
-
-// struct ClaritestBlock {
-// }
-
-// struct ClaritestChain {
-// }
-
-// struct ClaritestAccount {
-// }
-
-// impl ClaritestChain {
-
-//     fn new() -> ClaritestChain {
-//         ClaritestChain {
-//         }
-//     }
-
-//     fn test() {
-//         let config = ClarinetConfig::new();
-//         let chain = ClaritestChain::new(config);
-//         chain.start();
-
-//     }
-// }
-
-// #[claritest()]
-// fn test_box_btc(chain: ClaritestChain, accounts: Hashmap<String, ClaritestAccount>) {
-//     let block = chain.mine_block(vec![
-//         tx!("(contract-call? 'ST000000000000000000002AMW42H.bbtc create-box size fee)"),
-//     ]);
-
-//     let res = chain.read("(contract-call? 'ST000000000000000000002AMW42H.bbtc create-box size fee)");
-// }
